0731 GlobalException 처리 완료

+++ src/main/java/com/green/gajigaji/review/ReviewService.java
@@ -2,7 +2,7 @@ package com.green.gajigaji.review;
 
 import com.green.gajigaji.common.CheckMapper;
 import com.green.gajigaji.common.model.CustomFileUtils;
-import com.green.gajigaji.review.exception.CustomException;
+import com.green.gajigaji.common.exception.CustomException;
 import com.green.gajigaji.review.exception.ReviewErrorCode;
 import com.green.gajigaji.review.model.*;
 import lombok.RequiredArgsConstructor;
@@ -10,7 +10,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
-
 import java.util.List;
 
 @Service
@@ -23,7 +22,7 @@ public class ReviewService {
     private String path = "review/";
 
     @Transactional
-    public PostReviewRes postReview(List<MultipartFile> pics, PostReviewReq p) throws Exception{
+    public PostReviewRes postReview(List<MultipartFile> pics, PostReviewReq p){
         if(chkMapper.checkPostedReview(p.getReviewPlanSeq(), p.getReviewPlmemberSeq()) == 1) {
             throw new CustomException(ReviewErrorCode.DUPLICATED_REVIEW);
         }
@@ -33,7 +32,7 @@ public class ReviewService {
 
             if(result == 0) {
                 log.error("Review Post Result == 0");
-                throw new CustomException(ReviewErrorCode.UNIDENTIFIED_ERROR);
+                throw new CustomException(ReviewErrorCode.REVIEW_POST_ERROR);
             }
             if (pics == null) {
                 return PostReviewRes.builder()
@@ -48,16 +47,22 @@ public class ReviewService {
                     .pics(ppic.getFileNames())
                     .build();
         } catch (Exception e) {
-            throw new Exception();
+            throw new CustomException(ReviewErrorCode.UNIDENTIFIED_ERROR);
         }
     }
 
     @Transactional
-    public List<GetReviewPicDto> patchReview(List<MultipartFile> pics, PatchReviewReq p) throws Exception{
+    public List<GetReviewPicDto> patchReview(List<MultipartFile> pics, PatchReviewReq p){
         try {
-            mapper.deleteReviewPics(p.getReviewSeq());
+            if(chkMapper.checkReview(p.getReviewSeq()) == null) { // 리뷰 있는지 체크
+                throw new CustomException(ReviewErrorCode.NOT_FOUND_REVIEW);
+            }
+            if(chkMapper.checkReviewPics(p.getReviewSeq()) != null) { // 리뷰 사진 있으면 사진삭제+폴더삭제
+                mapper.deleteReviewPics(p.getReviewSeq());
+                customFileUtils.deleteFolder(path + p.getReviewSeq());
+            }
+            
             mapper.patchReview(p);
-            customFileUtils.deleteFolder(path + p.getReviewSeq());
 
             if (pics == null) {
                 return null;
@@ -67,19 +72,22 @@ public class ReviewService {
             return mapper.getPic(ppic.getReviewSeq());
         } catch(Exception e) {
             log.error("Review Patch Error");
-            throw new CustomException(ReviewErrorCode.UNIDENTIFIED_ERROR);
+            throw new CustomException(ReviewErrorCode.REVIEW_PATCH_ERROR);
         }
     }
 
     @Transactional
     public int deleteReview(long reviewSeq) {
         try {
+            if(chkMapper.checkReview(reviewSeq) == null) {
+                throw new CustomException(ReviewErrorCode.NOT_FOUND_REVIEW);
+            }
             mapper.deleteReviewPics(reviewSeq);
             mapper.deleteReviewFavs(reviewSeq);
             int result = mapper.deleteReview(reviewSeq);
             if(result == 0) {
                 log.error("Review Delete Result == 0");
-                throw new CustomException(ReviewErrorCode.UNIDENTIFIED_ERROR);
+                throw new CustomException(ReviewErrorCode.REVIEW_DELETE_ERROR);
             }
             return result;
         }catch (Exception e) {
@@ -127,13 +135,14 @@ public class ReviewService {
         }
     }
 
-    public PostReviewPicDto postPics(long reviewSeq, List<MultipartFile> pics, String path) throws Exception {
+    public PostReviewPicDto postPics(long reviewSeq, List<MultipartFile> pics, String path) {
         try {
             PostReviewPicDto ppic = PostReviewPicDto.builder()
                     .reviewSeq(reviewSeq)
                     .build();
 
             customFileUtils.makeFolders(path);
+
             for (MultipartFile pic : pics) {
                 String fileName = customFileUtils.makeRandomFileName(pic);
                 String target = String.format("%s/%s", path, fileName);
@@ -144,7 +153,7 @@ public class ReviewService {
 
             if (result == 0) {
                 log.error("Pics Post Result == 0");
-                throw new CustomException(ReviewErrorCode.UNIDENTIFIED_ERROR);
+                throw new CustomException(ReviewErrorCode.REVIEW_PICS_POST_ERROR);
             }
 
             return ppic;
@@ -156,6 +165,10 @@ public class ReviewService {
 
     public int toggleReviewFav(GetReviewFavToggleReq p) {
         try {
+            if(chkMapper.checkReview(p.getReviewFavReviewSeq()) == null) {
+                throw new CustomException(ReviewErrorCode.NOT_FOUND_REVIEW);
+            }
+
             int result = mapper.deleteReviewFav(p);
 
             if (result == 1) {
